# docs/swagger.yaml
basePath: /kitchencontrol/api/v1
definitions:
...
  usecase.OrderCreateInputDTO:
    properties:
      cpf:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemCreateInputDTO'
        type: array
    type: object
  usecase.OrderCreateOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  usecase.OrderFindAllOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  usecase.OrderFindByIdOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  usecase.OrderFindByParamOutputDTO:
    properties:
      created_at:
        type: string
      customercpf:
        type: string
      delivery_number:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/usecase.OrderItemDTO'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  usecase.OrderItemCreateInputDTO:
    properties:
      productid:
        type: string
      quantity:
        type: integer
    type: object
  usecase.OrderItemDTO:
    properties:
      id:
        type: string
      price:
        type: number
      productid:
        type: string
      quantity:
        type: integer
      status:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is fiap fase 2 challenge project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiap Fase 2 Challenge Clean Arch API - 9SOAT
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Retorna todos os pedidos (orders) registrados no sistema. Se não
        houver pedidos, retorna um erro (404).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.OrderFindAllOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Cria um peddo (order) no sistema. O cliente (customer) pode ou
        não de identificar. Se o cliente não se identificar, o pedido será registrado
        como anônimo. O produto, porém, deve ter sido previamente cadastrado.
      parameters:
      - description: cria nova Order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.OrderCreateInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.OrderCreateOutputDTO'
        "400":
          description: invalid data
          schema:
            type: string
        "409":
          description: Order already exists
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create Order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a Order and their items by id
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.OrderFindByIdOutputDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Order not found
          schema:
            type: string
      summary: Get a Order by id
      tags:
      - Orders
  /orders/confirmed:
    get:
      consumes:
      - application/json
      description: Retorna todos os pedidos (orders) registrados no sistema. Se não
        houver pedidos, retorna um erro (404).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.OrderFindByParamOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all confirmed orders
      tags:
      - Orders
  /orders/paid:
    get:
      consumes:
      - application/json
      description: Retorna todos os pedidos (orders) registrados no sistema que tenham
        status de pagamento confirmado. Se não houver pedidos, retorna um erro (404).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.OrderFindByParamOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all confirmed orders
      tags:
      - Orders
  /orders/pending:
    get:
      consumes:
      - application/json
      description: Retorna todos os pedidos (orders) registrados no sistema sem o
        pagamento confirmado. Se não houver pedidos, retorna um erro (404).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.OrderFindByParamOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all orders not confirmed
      tags:
      - Orders
...
swagger: "2.0"